backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
autopep8==2.3.2
backcall==0.2.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
decorator==5.1.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
executing==2.2.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
ipython==8.32.0
ipython-genutils==0.2.0
jedi==0.19.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pytoolconfig==1.3.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
rope==1.13.0
setuptools==75.8.0
simplegeneric==0.8.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
